# groupGifts-api-rest/src/main/resources/api.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /giver_contribution:
    post:
      tags:
        - giver_contribution-controller
      operationId: createGiver_contribution
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Giver_contributionDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Giver_contributionDto'
    put:
      tags:
        - giver_contribution-controller
      operationId: updateGiver_contribution
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Giver_contributionDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Giver_contributionDto'
        '400':
          description: Invalid Giver_contribution params
        '404':
          description: Giver_contribution not found
    delete:
      tags:
        - giver_contribution-controller
      operationId: removeGiver_contribution
      parameters:
        - name: giver_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: contribution_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
  /giver_contribution/filter:
    get:
      tags:
        - giver_contribution-controller
      operationId: findGiver_contribution
      parameters:
        - name: giver_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: contribution_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Giver_contributionDto'

# GIFT_GIVER
  /gift_giver:
    post:
      tags:
        - gift_giver-controller
      operationId: createGift_giver
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Gift_giverDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gift_giverDto'
    put:
      tags:
        - gift_giver-controller
      operationId: updateGift_giver
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Gift_giverDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gift_giverDto'
        '400':
          description: Invalid Gift_giver params
        '404':
          description: Gift_giver not found
    delete:
      tags:
        - gift_giver-controller
      operationId: removeGift_giver
      parameters:
        - name: gift_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: giver_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
  /gift_giver/filter:
    get:
      tags:
        - gift_giver-controller
      operationId: findGift_giver
      parameters:
        - name: giver_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: gift_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gift_giverDto'
components:
  schemas:
...
    Giver_contributionDto:
      required:
        - giver_id
        - contribution_id
      type: object
      properties:
        giver_id:
          type: integer
          format: int64
        contribution_id:
          type: integer
          format: int64

# GIFT_GIVER
    Gift_giverDto:
      required:
        - gift_id
        - giver_id
      type: object
      properties:
        gift_id:
          type: integer
          format: int64
        giver_id:
          type: integer
          format: int64