# groupGifts-api-rest/target/generated-sources/openapi/src/main/java/DAW/proyecto/groupGifts/rest/GiverContributionApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package DAW.proyecto.groupGifts.rest;

import DAW.proyecto.groupGifts.dto.GiverContributionDto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-11-11T20:19:36.138872500+01:00[Europe/Madrid]")
@Validated
@Tag(name = "giver_contribution-controller", description = "the giver_contribution-controller API")
public interface GiverContributionApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /giver_contribution
     *
     * @param giverContributionDto  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "createGiverContribution",
        tags = { "giver_contribution-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GiverContributionDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/giver_contribution",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<GiverContributionDto> createGiverContribution(
        @Parameter(name = "GiverContributionDto", description = "", required = true) @Valid @RequestBody GiverContributionDto giverContributionDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"contribution_id\" : 6, \"giver_id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /giver_contribution/filter
     *
     * @param giverId  (required)
     * @param contributionId  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "findGiverContribution",
        tags = { "giver_contribution-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = GiverContributionDto.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/giver_contribution/filter",
        produces = { "application/json" }
    )
    default ResponseEntity<List<GiverContributionDto>> findGiverContribution(
        @NotNull @Parameter(name = "giver_id", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "giver_id", required = true) Long giverId,
        @NotNull @Parameter(name = "contribution_id", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "contribution_id", required = true) Long contributionId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"contribution_id\" : 6, \"giver_id\" : 0 }, { \"contribution_id\" : 6, \"giver_id\" : 0 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /giver_contribution
     *
     * @param giverId  (required)
     * @param contributionId  (required)
     * @return No Content (status code 204)
     */
    @Operation(
        operationId = "removeGiverContribution",
        tags = { "giver_contribution-controller" },
        responses = {
            @ApiResponse(responseCode = "204", description = "No Content")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/giver_contribution"
    )
    default ResponseEntity<Void> removeGiverContribution(
        @Parameter(name = "giver_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("giver_id") Long giverId,
        @Parameter(name = "contribution_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("contribution_id") Long contributionId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /giver_contribution
     *
     * @param giverContributionDto  (required)
     * @return OK (status code 200)
     *         or Invalid Giver_contribution params (status code 400)
     *         or Giver_contribution not found (status code 404)
     */
    @Operation(
        operationId = "updateGiverContribution",
        tags = { "giver_contribution-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GiverContributionDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid Giver_contribution params"),
            @ApiResponse(responseCode = "404", description = "Giver_contribution not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/giver_contribution",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<GiverContributionDto> updateGiverContribution(
        @Parameter(name = "GiverContributionDto", description = "", required = true) @Valid @RequestBody GiverContributionDto giverContributionDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"contribution_id\" : 6, \"giver_id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}